% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Create a Cleveland Dot Chart of the correlation}
\usage{
\method{plot}{ryx}(x)
}
\arguments{
\item{x}{The list result from the ryx function}
}
\value{
A Cleveland Dot Chart of the correlation
}
\description{
Create a Cleveland Dot Chart of the correlation
}
\examples{
\dontrun{
library(MASS)
x <- ryx(Boston, y="medv")
plot.ryx(x)
}
}
