% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxy2-package.R
\docType{package}
\name{rxy2-package}
\alias{rxy2}
\alias{rxy2-package}
\title{rxy2: Correlation Table, Print, Summary, and Plot}
\description{
This package takes (dataset, y, x) in the ryx function to return a list of the y variable, the x variable, and a correlation table where y is correlated with every variable in x. The package contains the functions ryx(data, y,x) print(x) summary(x) plot(x) The input for print, summary and plot should be the results from the ryx function.
}
\author{
\strong{Maintainer}: Calvin Gao \email{cgao@wesleyan.edu}

}
\keyword{internal}
