% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Get a list of input variables y and x and a correlation table of y with all of x}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{The dataset that the variables are from}

\item{y}{The variable used to correlate with each of x}

\item{x}{The variable/variables to correlate with y}
}
\value{
A list of 1. y 2. x 3. correlation table of y correlated with each of x
}
\description{
This \code{ryx} returns a list that can be used with other functions in this package.
}
\examples{
\dontrun{
# Correlation of "medv" in the Boston dataset with every other variable in the dataset
library(MASS)
x <- ryx(Boston, y="medv")
x
}
}
