% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summary of the correlation results}
\usage{
\method{summary}{ryx}(x, digits = 3)
}
\arguments{
\item{x}{The list result from the ryx function}

\item{digits}{The number of digits we want to see in the numbers within the summary}
}
\value{
A paragraph summarizing the results of the correlation
}
\description{
Summary of the correlation results
}
\examples{
\dontrun{
library(MASS)
x <- ryx(Boston, y="medv")
summary.ryx(x, digits = 3)
}
}
